[REQ] #55639 Изменены типы числовых свойств
Свойства типа количества колонок, размеров формы и другие количественные показатели
не могут быть отрицательными. Также их диапазон скорее всего не превысит диапазона
`ushort`.

+++ src/OFP/ObjectModel/FormElements/Data/TableBox.cs
@@ -78,7 +78,7 @@ namespace OFP.ObjectModel.FormElements.Data
         /// <summary>
         /// ВысотаШапки.
         /// </summary>
-        public int HeaderHeight { get; set; }
+        public ushort HeaderHeight { get; set; }
 
         /// <summary>
         /// ВысотаПодвала.
@@ -98,12 +98,12 @@ namespace OFP.ObjectModel.FormElements.Data
         /// <summary>
         /// ФиксацияСлева.
         /// </summary>
-        public int CountOfColumnsFixedLeft { get; set; }
+        public ushort CountOfColumnsFixedLeft { get; set; }
 
         /// <summary>
         /// ФиксацияСправа.
         /// </summary>
-        public int CountOfColumnsFixedRight { get; set; }
+        public ushort CountOfColumnsFixedRight { get; set; }
 
         /// <summary>
         /// АвтоВводНовойСтроки.
